Adjust query to select all 3 columns
This fixes a bug where the single selected column couldn't be unpacked
in the list comprehension below.

Although columns apart from C_NAME aren't being used at the moment,
it is better for the query to be generic, and only the Python code
to change should different behaviour be desired in the future.